requirements.txt
chainlit>=1.1.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=1.1.0
supabase>=2.0.0
pypdf>=4.0.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
httpx>=0.27.0